Module Log.Reader

val pipe : [< Output.Format.machine_readable ] -> string -> Message.t Async_kernel.Pipe.Reader.t

pipe format filename returns a pipe of all the messages in the log. Errors encountered when opening or reading the file will be thrown as exceptions into the monitor current at the time pipe is called.

val pipe_of_reader : [< Output.Format.machine_readable ] -> Reader.t -> Message.t Async_kernel.Pipe.Reader.t

module Expert : sig ... end