Module Process.Lines_or_sexp

Lines_or_sexp is useful for rendering a string nicely in a sexp, avoiding quoting if the string is multi-line or was produced by converting a sexp to a string. Output.sexp_of_t uses Lines_or_sexp to nicely render stdout and stderr of a child process.

type t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val create : string -> t