Module Reader.Read_result

type 'a t = [

| `Ok of 'a

| `Eof

]

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

val __t_of_sexp__ : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

include Core.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'a t) -> 'a t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'a t

map t ~f is t >>| f.

val join : 'a t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> 'a t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a t

val all_unit : 'a t list -> 'a t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end