Module Private.Very_low_priority_work

module Worker_result : sig ... end

val enqueue : f:(unit -> Worker_result.t) -> unit

Enqueue some low-priority work to be done. The work will happen at some point, but Async will choose when is the best time to do it. f will be called until it returns Finished.