Module Require_explicit_time_source.Date

include module type of struct include Core.Date end

type t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include Core.Hashable.S_binable with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core.Hashtbl.Hashable.t

module Table : Core.Hashtbl.S_binable with type key = t

module Hash_set : Core.Hash_set.S_binable with type elt = t

module Hash_queue : Core.Hash_queue.S with type key = t

converts a string to a date in the following formats:

m/d/y

y-m-d (valid iso8601_extended)

DD MMM YYYY

DDMMMYYYY

YYYYMMDD

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Core.Comparable.S_binable with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

include Core.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : ( t, comparator_witness ) Core.Comparator.comparator

module Map : Core.Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core.Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

val create_exn : y:Base.Int.t -> m:Core.Month.t -> d:Base.Int.t -> t

create_exn ~y ~m ~d creates the date specified in the arguments. Arguments are validated, and are not normalized in any way. So, days must be within the limits for the month in question, numbers cannot be negative, years must be fully specified, etc.

val of_string_iso8601_basic : Base.String.t -> pos:Base.Int.t -> t

For details on this ISO format, see:

http://www.wikipedia.org/wiki/iso8601

val to_string_iso8601_basic : t -> Base.String.t

YYYYMMDD

val to_string_american : t -> Base.String.t

MM/DD/YYYY

val day : t -> Base.Int.t

val month : t -> Core.Month.t

val year : t -> Base.Int.t

val day_of_week : t -> Core.Day_of_week.t

Only accurate after 1752-09

val week_number_and_year : t -> Base.Int.t * Base.Int.t

Week of the year, from 1 to 53, along with the week-numbering year to which the week belongs. The week-numbering year may not correspond to the calendar year in which the provided date occurs.

According to ISO 8601, weeks start on Monday, and the first week of a year is the week that contains the first Thursday of the year. This means that dates near the end of the calendar year can have week number 1 and belong to the following week-numbering year, and dates near the beginning of the calendar year can have week number 52 or 53 and belong to the previous week-numbering year.

The triple (week-numbering year, week number, week day) uniquely identifies a particular date, which is not true if the calendar year is used instead.

val week_number : t -> Base.Int.t

See week_number_and_year for the meaning of week number.

val is_weekend : t -> Base.Bool.t

val is_weekday : t -> Base.Bool.t

val is_business_day : t -> is_holiday:( t -> Base.Bool.t ) -> Base.Bool.t

Monday through Friday are business days, unless they're a holiday.

val add_days : t -> Base.Int.t -> t

add_days t n adds n days to t and returns the resulting date.

Inaccurate when crossing 1752-09.

val add_months : t -> Base.Int.t -> t

add_months t n returns date with max days for the month if the date would be invalid. e.g. adding 1 month to Jan 30 results in Feb 28 due to Feb 30 being an invalid date, Feb 29 is returned in cases of leap year.

In particular, this means adding x months and then adding y months isn't the same as adding x + y months, and in particular adding x months and then -x months won't always get you back where you were. *

val add_years : t -> Base.Int.t -> t

add_years t n has the same semantics as add_months for adding years to Feb 29 of a leap year, i.e., when the addition results in a date in a non-leap year, the result will be Feb 28 of that year.

val diff : t -> t -> Base.Int.t

diff t1 t2 returns date t1 minus date t2 in days.

val diff_weekdays : t -> t -> Base.Int.t

diff_weekdays t1 t2 returns the number of weekdays in the half-open interval [t2,t1) if t1 >= t2, and - diff_weekdays t2 t1 otherwise.

val diff_weekend_days : t -> t -> Base.Int.t

diff_weekend_days t1 t2 returns the number of days that are weekend days in the half-open interval [t2,t1) if t1 >= t2, and - diff_weekend_days t2 t1 otherwise.

val add_weekdays_rounding_backward : t -> Base.Int.t -> t

First rounds the given date backward to the previous weekday, if it is not already a weekday. Then advances by the given number of weekdays, which may be negative.

val add_weekdays_rounding_forward : t -> Base.Int.t -> t

First rounds the given date forward to the next weekday, if it is not already a weekday. Then advances by the given number of weekdays, which may be negative.

val add_business_days_rounding_backward : t -> is_holiday:( t -> Base.Bool.t ) -> Base.Int.t -> t

First rounds the given date backward to the previous business day, i.e. weekday not satisfying is_holiday, if it is not already a business day. Then advances by the given number of business days, which may be negative.

val add_business_days_rounding_forward : t -> is_holiday:( t -> Base.Bool.t ) -> Base.Int.t -> t

First rounds the given date forward to the next business day, i.e. weekday not satisfying is_holiday, if it is not already a business day. Then advances by the given number of business days, which may be negative.

val add_weekdays : t -> Base.Int.t -> t

add_weekdays t 0 returns the next weekday if t is a weekend and t otherwise. Unlike add_days this is done by looping over the count of days to be added (forward or backwards based on the sign), and is O(n) in the number of days to add. Beware, add_weekdays sat 1 or add_weekdays sun 1 both return the next tue, not the next mon. You may want to use following_weekday if you want the next following weekday, following_weekday (fri|sat|sun) would all return the next mon.

val add_weekdays_rounding_in_direction_of_step : t -> Base.Int.t -> t

val add_business_days : t -> is_holiday:( t -> Base.Bool.t ) -> Base.Int.t -> t

add_business_days t ~is_holiday n returns a business day even when n=0. add_business_days ~is_holiday:(fun _ -> false) ... is the same as add_weekdays.

If you don't want to skip Saturday or Sunday, use add_days_skipping.

val add_business_days_rounding_in_direction_of_step : t -> is_holiday:( t -> Base.Bool.t ) -> Base.Int.t -> t

val add_days_skipping : t -> skip:( t -> Base.Bool.t ) -> Base.Int.t -> t

add_days_skipping t ~skip n adds n days to t, ignoring any date satisfying skip, starting at the first date at or after t that does not satisfy skip. For example, if skip t = true, then add_days_skipping t ~skip 0 > t.

add_business_days and add_weekdays are special cases of add_days_skipping.

val dates_between : min:t -> max:t -> t Base.List.t

the following returns a closed interval (endpoints included)

val business_dates_between : min:t -> max:t -> is_holiday:( t -> Base.Bool.t ) -> t Base.List.t

val weekdays_between : min:t -> max:t -> t Base.List.t

val previous_weekday : t -> t

val following_weekday : t -> t

val first_strictly_after : t -> on:Core.Day_of_week.t -> t

first_strictly_after t ~on:day_of_week returns the first occurrence of day_of_week strictly after t.

val days_in_month : year:Base.Int.t -> month:Core.Month.t -> Base.Int.t

days_in_month ~year ~month returns the number of days in month, using year only if month = Month.Feb to check if there is a leap year.

Incorrect for September 1752.

val is_leap_year : year:Base.Int.t -> Base.Bool.t

is_leap_year ~year returns true if year is considered a leap year

val unix_epoch : t

The starting date of the UNIX epoch: 1970-01-01

gen generates dates between 1900-01-01 and 2100-01-01.

include Core.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Core.Quickcheck.Generator.t

gen_incl d1 d2 generates dates in the range between d1 and d2, inclusive, with the endpoints having higher weight than the rest. Raises if d1 > d2.

val gen_uniform_incl : t -> t -> t Core.Quickcheck.Generator.t

gen_uniform_incl d1 d2 generates dates chosen uniformly in the range between d1 and d2, inclusive. Raises if d1 > d2.

module Days : sig ... end

Days provides a linear representation of dates that is optimized for arithmetic on the number of days between dates, rather than for representing year/month/day components. This module is intended for use only in performance-sensitive contexts where dates are manipulated more often than they are constructed or deconstructed; most clients should use the ordinary t.

module Option : sig ... end

module Stable : sig ... end

module O : sig ... end

module Private : sig ... end

val of_time : Base.Float.t -> zone:Core__Zone.t -> t

Deprecations

val format : [ `Use_Date_unix ]

val of_tm : [ `Use_Date_unix ]

val parse : [ `Use_Date_unix ]

val today : zone:{Time}2.Zone.t -> t