Module Require_explicit_time_source.Time

deprecated [since 2022-04] Use [Time_float]

include module type of struct include Time_float_unix end

include module type of struct include Time_float_unix end

include module type of struct include Core.Time_float end

module Span = Time_float_unix.Span

module Zone = Time_float_unix.Zone

type t = Base.Float.t

A fully qualified point in time, independent of timezone.

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

include sig ... end

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_hash_lib.Hashable.S_any with type t := t

include Sexplib0.Sexpable.S with type t := t

include Sexplib0.Sexpable.Of_sexp with type t := t

include Sexplib0.Sexpable.Sexp_of with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t @@ portable

include Typerep_lib.Typerepable.S with type t := t

type underlying = Base.Float.t

include Bin_prot.Binable.S__local with type t := t

include Bin_prot.Binable.S_only_functions__local with type t := t

include sig ... end

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_hash_lib.Hashable.S_any with type t := t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t @@ portable

val typename_of_t : t Typerep_lib.Typename.t @@ portable

module Replace_polymorphic_compare = Time_float_unix.Replace_polymorphic_compare

include Base.Comparable.Comparisons__local with type t := t

include Base.Comparisons.Infix with type t := t

include Core.Interfaces.Robustly_comparable with type t := t

val next : t -> t

next t returns the next t (forwards in time)

val prev : t -> t

prev t returns the previous t (backwards in time)

val to_span_since_epoch : t -> Span.t

val of_span_since_epoch : Span.t -> t

include Base.Comparable.S__local__portable with type t := t

include Base.Comparisons.S__local with type t := t

include Base.Comparisons.Infix with type t := t

include Base.Comparator.S__portable with type t := t

type comparator_witness

module Date_and_ofday = Time_float_unix.Date_and_ofday

Equivalent to a Date.t and an Ofday.t with no time zone. A Date_and_ofday.t does not correspond to a single, unambiguous point in time.

val next_multiple : ?can_equal_after:Base.Bool.t @ local -> (base:t -> (after:t -> (interval:Span.t -> (Base.Unit.t -> t) @ local) @ local) @ local) @ local @@ portable

next_multiple ~base ~after ~interval returns the smallest time of the form:

  time = base + (k * interval)

where k >= 0 and time > after. It is an error if interval <= 0.

Supplying ~can_equal_after:true allows the result to satisfy time >= after.

val prev_multiple : ?can_equal_before:Base.Bool.t -> base:t -> before:t -> interval:Span.t -> Base.Unit.t -> t @@ portable

prev_multiple ~base ~before ~interval returns the largest time of the form:

  time = base + (k * interval)

where k >= 0 and time < before. It is an error if interval <= 0.

Supplying ~can_equal_before:true allows the result to satisfy time <= before.

Basic operations on times

val add : t -> Span.t -> t

add t s adds the span s to time t and returns the resulting time.

NOTE: adding spans as a means of adding days is not accurate, and may run into trouble due to shifts in daylight savings time, float arithmetic issues, and leap seconds. See the comment at the top of Zone.mli for a more complete discussion of some of the issues of time-keeping. For spans that cross date boundaries, use date functions instead.

val sub : t -> Span.t -> t

sub t s subtracts the span s from time t and returns the resulting time. See important note for add.

val diff : t -> t -> Span.t

diff t1 t2 returns time t1 minus time t2.

val abs_diff : t -> t -> Span.t

abs_diff t1 t2 returns the absolute span of time t1 minus time t2.

include Core.Quickcheck.S_range with type t := t

include Core.Quickcheck_intf.Definitions.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

Comparisons

val is_earlier : t -> than:t -> Base.Bool.t

val is_later : t -> than:t -> Base.Bool.t

Conversions

val of_date_ofday : ?prefer:Core__.Zone_intf.Earlier_or_later.t -> zone:Core__.Zone.t -> Core__.Date0.t -> Core__.Time_float0.Ofday.underlying -> t

Produces the absolute time of a date and time of day in this zone. When there are two occurrences, the result is determined by prefer.

val of_date_ofday_precise : Core__.Date0.t -> Core__.Time_float0.Ofday.underlying -> zone:Core__.Zone.t -> [ `Once of t | `Twice of t * t | `Never of t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : t -> zone:Core__.Zone.t -> Core__.Date0.t * Core__.Time_float0.Ofday.underlying

val to_date_ofday_precise : t -> zone:Core__.Zone.t -> Core__.Date0.t * Core__.Time_float0.Ofday.underlying * [ `Only | `Also_at of t | `Also_skipped of Core__.Date0.t * Core__.Time_float0.Ofday.underlying ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

- `Only         -> there is a one-to-one mapping between [t]'s and
                   [Date.t * Ofday.t] pairs
- `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                   (this date/time pair happened twice because the clock fell back)
- `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                   to a jump forward) that [of_date_ofday] would map to the same
                   [t].

val to_date : t -> zone:Core__.Zone.t -> Core__.Date0.t

val to_ofday : t -> zone:Core__.Zone.t -> Core__.Time_float0.Ofday.underlying

val reset_date_cache : Base.Unit.t -> Base.Unit.t

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

Unlike Time_ns, this module purposely omits max_value and min_value: 1. They produce unintuitive corner cases because most people's mental models of time do not include +/- infinity as concrete values 2. In practice, when people ask for these values, it is for questionable uses, e.g., as null values to use in place of explicit options.

val epoch : t

midnight, Jan 1, 1970 in UTC

val convert : from_tz:Core__.Zone.t -> to_tz:Core__.Zone.t -> Core__.Date0.t -> Core__.Time_float0.Ofday.underlying -> Core__.Date0.t * Core__.Time_float0.Ofday.underlying

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : t -> zone:Core__.Zone.t -> Span.t

Other string conversions

to_string uses the local time zone. of_string defaults to the local time zone if not specified, and is generous in what it will read in.

include Core.Interfaces.Stringable with type t := t

val to_filename_string : t -> zone:Core__.Zone.t -> Base.String.t

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Base.String.t -> zone:Core__.Zone.t -> t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : t -> zone:Core__.Zone.t -> Base.String.t

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : t -> zone:Core__.Zone.t -> Base.String.t

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : t -> zone:Core__.Zone.t -> Base.String.t Base.List.t

val to_string_trimmed : t -> zone:Core__.Zone.t -> Base.String.t

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : t -> zone:Core__.Zone.t -> Base.String.t

Same as to_string_abs, but without milliseconds and the timezone. May raise if zone offsets move the apparent value beyond min_value_representable and max_value_representable.

val to_sec_string_with_zone : t -> zone:Core__.Zone.t -> Base.String.t

Same as to_sec_string but includes timezone

val of_localized_string : zone:Core__.Zone.t -> Base.String.t -> t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val to_string_iso8601_basic : t -> zone:Core__.Zone.t -> Base.String.t

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> t -> ofday:Core__.Time_float0.Ofday.underlying -> zone:Core__.Zone.t -> t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.

val of_string_with_utc_offset : Base.String.t -> t @@ portable

of_string_with_utc_offset requires its input to have an explicit UTC offset, e.g. 2000-01-01 12:34:56.789012-23, or use the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.

val to_string_utc : t -> Base.String.t @@ portable

to_string_utc generates a time string with the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.

val arg_type : t Command.Arg_type.t @@ portable

String conversions use the local timezone by default. Sexp conversions use get_sexp_zone () by default, which can be overridden by calling set_sexp_zone. These default time zones are used when writing a time, and when reading a time with no explicit zone or UTC offset.

Sexps and strings display the date, ofday, and UTC offset of t relative to the appropriate time zone.

include Core.Identifiable.S__local with type t := t and type comparator_witness := comparator_witness and module Replace_polymorphic_compare := Replace_polymorphic_compare

include Bin_prot.Binable.S__local with type t := t

include Bin_prot.Binable.S_only_functions__local with type t := t

include sig ... end

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : t Bin_prot.Read.vtag_reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_hash_lib.Hashable.S_any with type t := t

include Sexplib0.Sexpable.S with type t := t

include Sexplib0.Sexpable.Of_sexp with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

include Sexplib0.Sexpable.Sexp_of with type t := t

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_hash_lib.Hashable.S_any with type t := t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Base.Comparable.S__local with type t := t with type comparator_witness := comparator_witness

include Base.Comparisons.S__local with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t with type comparator_witness := comparator_witness

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

include Core.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.T.comparator

module Map = Time_float_unix.Map

module Set = Time_float_unix.Set

include Core.Hashable.S_binable with type t := t

include Ppx_hash_lib.Hashable.S_any with type t := t

val hash_fold_t : t Ppx_hash_lib.hash_fold

val hash : t -> Ppx_hash_lib.Std.Hash.hash_value

val hashable : t Base.Hashable.t

module Table = Time_float_unix.Table

module Hash_set = Time_float_unix.Hash_set

module Hash_queue = Time_float_unix.Hash_queue

val get_sexp_zone : Base.Unit.t -> Zone.t @@ portable

val set_sexp_zone : Zone.t -> Base.Unit.t @@ portable

include Core.Interfaces.Robustly_comparable with type t := t

val (>=.) : t @ local -> (t @ local -> bool) @ local

val (<=.) : t @ local -> (t @ local -> bool) @ local

val (=.) : t @ local -> (t @ local -> bool) @ local

val (>.) : t @ local -> (t @ local -> bool) @ local

val (<.) : t @ local -> (t @ local -> bool) @ local

val (<>.) : t @ local -> (t @ local -> bool) @ local

val robustly_compare : t @ local -> (t @ local -> int) @ local

val of_date_ofday_zoned : Core__.Date0.t -> Core.Time_float.Ofday.Zoned.t -> t @@ portable

Conversion functions that involved Ofday.Zoned.t, exactly analogous to the conversion functions that involve Ofday.t

val to_date_ofday_zoned : t -> zone:Zone.t -> Core__.Date0.t * Core.Time_float.Ofday.Zoned.t @@ portable

val to_ofday_zoned : t -> zone:Zone.t -> Core.Time_float.Ofday.Zoned.t @@ portable

val to_string_fix_proto : [ `Utc | `Local ] -> t -> Base.String.t @@ portable

val of_string_fix_proto : [ `Utc | `Local ] -> Base.String.t -> t @@ portable

val of_string_abs : Base.String.t -> t @@ portable

This is like of_string except that if the string doesn't specify the zone then it raises rather than assume the local timezone.

val of_string_gen : if_no_timezone:[ `Fail | `Local | `Use_this_one of Zone.t ] -> Base.String.t -> t @@ portable

of_string_gen ~if_no_timezone s attempts to parse s to a t. If s doesn't supply a time zone if_no_timezone is consulted.

val t_of_sexp_abs : Core.Sexp.t -> t @@ portable

t_of_sexp_abs sexp as t_of_sexp, but demands that sexp indicate the timezone the time is expressed in.

val sexp_of_t_abs : t -> zone:Zone.t -> Core.Sexp.t @@ portable

module Exposed_for_tests = Time_float_unix.Exposed_for_tests

include Diffable.S_atomic with type t := t

module Diff = Time_float_unix.Diff

module Stable = Time_float_unix.Stable

Notes on time

This library replicates and extends the functionality of the standard Unix time handling functions (currently exposed in the Unix module, and indirectly through the Time module).

Things you should know before delving into the mess of time...

Some general resources (summarized information also appears below)

general overview   - http://www.twinsun.com/tz/tz-link.htm
zone abbreviations - http://blogs.msdn.com/oldnewthing/archive/2008/03/07/8080060.aspx
leap seconds       - http://en.wikipedia.org/wiki/Leap_second
epoch time         - http://en.wikipedia.org/wiki/Unix_time
UTC/GMT time       - http://www.apparent-wind.com/gmt-explained.html
TAI time           - http://en.wikipedia.org/wiki/International_Atomic_Time
Almost every possible time measurement -
  http://www.ucolick.org/~sla/leapsecs/timescales.html

Standards for measuring time

Epoch time/Unix time/Posix time: Defined as the number of seconds that have passed since midnight, January 1st, 1970 GMT. However, under epoch time, a day is always 86,400 seconds long, and a minute never contains more than 60 total seconds. In other words, epoch time does not take leap seconds into account properly. What a POSIX compliant system does during a leap second depends on the way in which its clock is managed. It either ignores it, replays the second, or causes a second to last longer than a second (retards the second). The important thing to remember is that however the transition is managed, all days start on an evenly divisible multiple of 86,400.

GMT/Greenwich Mean Time/Greenwich Civil Time: The time based on the movement of the sun relative to the meridian through the Old Greenwich Observatory (0 degrees). The movement of the sun in this case is a "mean" movement of the sun to adjust for slight eccentricities in the rotation of the earth, as well as for the effect of the tilt of the earth on the visible speed of the sun across the sky at different times of the year. GMT is often used synonymously with the term UTC (see below), but may also be used to refer to the time system described here, which differs from UTC (as of 2009) by ~1 second.

Standard Time: The time based on the adjusted (as in GMT) movement of the sun over a point on the earth that is not Greenwich. Colloquially, the time in a time zone without accounting for any form of daylight savings time.

Wall Clock Time: The time as it appears on a clock on the wall in a given time zone. Essentially this is standard time with DST adjustments.

TAI: International atomic time. The time based on a weighted average of the time kept by roughly 300 atomic clocks worldwide. TAI is written using the same format as normal solar (also called civil) times, but is not based on, or adjusted for the apparent solar time. Thus, as of 2009 TAI appears to be ahead of most other time systems by ~34 seconds when written out in date/time form (2004-09-17T00:00:32 TAI is 2004-09-17T00:00:00 UTC)

UTC/Universal Coordinated Time: Often taken as just another term for GMT, UTC is actually TAI adjusted with leap seconds to keep it in line with apparent solar time. Each UTC day is not an exact number of seconds long (unlike TAI or epoch time), and every second is exactly one real second long (unlike GMT, which is based entirely on the apparent motion of the sun, meaning that seconds under GMT slowly get longer as the earth's rotation slows down). Leap seconds are determined by the rotation of the earth, which is carefully measured by the International Earth Rotation Service in Paris, France using a combination of satellite and lunar laser ranging, very long baseline interferometry, and Navstar Global Positioning System (GPS) stations. This isn't important for using UTC, but is very cool. UTC is not well defined before about 1960.

Windows File Time: The number of 100-nanosecond intervals that have elapsed since 12:00 A.M. January 1, 1601, UTC. This is great because UTC has no meaning in 1601 (being based on atomic timekeeping technologies that didn't exist then), and also because 1601 predates the development of even reasonably accurate clocks of any sort. The reasoning behind the Windows epoch time choice is that "The Gregorian calendar operates on a 400-year cycle, and 1601 is the first year of the cycle that was active at the time Windows NT was being designed. In other words, it was chosen to make the math come out nicely." (http://blogs.msdn.com/oldnewthing/archive/2009/03/06/9461176.aspx)

VBScript (this is my favorite): http://blogs.msdn.com/ericlippert/archive/2003/09/16/eric-s-complete-guide-to-vt-date.aspx

All of these systems start to exhibit problems as you go further back in time, partly because truly accurate timekeeping didn't make an appearance until roughly 1958, and partly because different parts of the world didn't actually have well defined time zones for a long time. If you go back far enough, you run into the switch between the Julian (old) and the Gregorian calendar, which happened at different times in history in different places in the world.

How does a system determine what time zone it is in?

Check to see if the TZ environment variable is set. If it is, it can be set to one of three forms, two of which are rarely, if ever used see:

http://www.opengroup.org/onlinepubs/000095399/basedefs/xbd_chap08.html

for more information on the obscure forms. The common form represents a relative path from the base /usr/share/zoneinfo/posix, and is generally in the form of a continent or country name paired with a city name (Europe/London, America/New_York). This is used to load the specified file from disk, which contains a time zone database in zic format (man tzfile).

If TZ is not set, the system will try to read the file located at /etc/localtime, which must be a zic timezone database (and which is often just a symlink into /usr/share/zoneinfo/posix).

If /etc/localtime cannot be found, then the system is assumed to be in GMT.

It's worth noting that under this system there is no place on the system to go to get the name of the file you are using (/etc/localtime may not be a link, and may just be a copy, or its own database not represented in /usr/share/zoneinfo). Additionally, the names of the files in the system zoneinfo database follow an internal standard, and there is no established standard for naming timezones. So even if you were using one of these files, and you did know its name, you cannot assume that that name matches any timezone specified by any other system or description.

One common misconception about time zones is that the standard time zone abbreviations can be used. For instance, EST surely refers to Eastern Standard Time. This is unfortunately not true - CST can refer to China Central Time, Central Standard Time, or Cuba Summer Time for instance - and time zone libraries that appear to correctly parse times that use time zone abbreviations do so by using a heuristic that usually assumes you mean a time in the US or Europe, in that order. Time zones also sometimes use two different abbreviations depending on whether the time in question is in standard time, or daylight savings time. These abbreviations are kept in the timezone databases, which is how programs like date manage to output meaningful abbreviations. The only poorly specified operation is reading in times with abbreviations.

This library contains a function that attempts to make an accurate determination of the machine timezone by testing the md5 sum of the currently referenced timezone file against all of the possible candidates in the system database. It additionally makes some adjustments to return the more common timezone names since some files in the database are duplicated under several names. It returns an option because of the problems mentioned above.

The problems with string time conversions

There are two cases where string time conversions are problematic, both related to daylight savings time.

In the case where time jumps forward one hour, there are possible representations of times that never happened 2006-04-02T02:30:00 in the eastern U.S. never happened for instance, because the clock jumped forward one hour directly from 2 to 3. Unix time zone libraries asked to convert one of these times will generally produce the epoch time that represents the time 1/2 hour after 2 am, which when converted back to a string representation will be T03:30:00.

The second case is when the clocks are set back one hour, which causes one hour of time to happen twice. Converting a string in this range without further specification into an epoch time is indeterminate since it could be referring to either of two times. Unix libraries handle this by either allowing you to pass in a dst flag to the conversion function to specify which time you mean, or by using a heuristic to guess which time you meant.

The existence of both cases make a strong argument for serializing all times in UTC, which doesn't suffer from these issues.

val of_tm : Core_unix.tm -> zone:Zone.t -> t @@ portable

of_tm converts a Unix.tm (mirroring a struct tm from the C stdlib) into a Time.t. Note that the tm_wday, tm_yday, and tm_isdst fields are ignored.

val pause : Span.t -> unit @@ portable

pause span sleeps for span time.

val interruptible_pause : Span.t -> [ `Ok | `Remaining of Span.t ] @@ portable

interruptible_pause span sleeps for span time unless interrupted (e.g. by delivery of a signal), in which case the remaining unslept portion of time is returned.

val pause_forever : unit -> Core.never_returns @@ portable

pause_forever sleeps indefinitely.

val format : t -> string -> zone:Zone.t -> string

format t fmt formats the given time according to fmt, which follows the formatting rules given in 'man strftime'. The time is output in the given timezone. Here are some commonly used control codes:

  %Y - year (4 digits)
  %y - year (2 digits)
  %m - month
  %d - day
  %H - hour
  %M - minute
  %S - second

a common choice would be: %Y-%m-%d %H:%M:%S

Although %Z and %z are interpreted as format strings, neither are correct in the current implementation. %Z always refers to the local machine timezone, and does not correctly detect whether DST is active. The effective local timezone can be controlled by setting the "TZ" environment variable before calling format. %z behaves unreliably and should be avoided.

Not all strftime control codes are standard; the supported subset will depend on the C libraries linked into a given executable.

val format_with_locale : t -> string -> zone:Zone.t -> locale:Core_unix.Locale.t -> string @@ portable

A thread-safe version of format where you are obliged to pass locale (though you should not free the locale while the call is in progress). The syntax of the format string is the same.

val parse : ?allow_trailing_input:bool -> string -> fmt:string -> zone:Zone.t -> t

parse string ~fmt ~zone parses string, according to fmt, which follows the formatting rules given in 'man strptime'. The time is assumed to be in the given timezone.

  %Y - year (4 digits)
  %y - year (2 digits)
  %m - month
  %d - day
  %H - hour
  %M - minute
  %S - second

Raise if allow_trailing_input is false and fmt does not consume all of the input.

val parse_with_locale : ?allow_trailing_input:bool -> string -> fmt:string -> zone:Zone.t -> locale:Core_unix.Locale.t -> t @@ portable

A thread-safe version of parse where you are obliged to pass locale (though you should not free the locale while the call is in progress). The syntax of the format string is the same.

module Ofday : sig ... end

val now : unit -> t

deprecated [since 2016-02] Use [Time_source]