Module Ofday.Zoned

type t

Sexps look like "(12:01 nyc)"

Two t's may or may not correspond to the same times depending on which date they're evaluated.

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

Strings look like "12:01 nyc"

val of_string : string -> t

val to_string : t -> string

val arg_type : t Core_kernel.Command.Arg_type.t

val create : Base.Float.t -> Zone.t -> t

val create_local : Base.Float.t -> t

val ofday : t -> Base.Float.t

val zone : t -> Zone.t

val to_time : t -> Core_kernel.Date.t -> Base.Float.t

module With_nonchronological_compare : sig ... end