Module Time.Stable

module V1 : sig ... end

module With_utc_sexp : sig ... end

Provides a sexp representation that is independent of the time zone of the machine writing it. V1.t_of_sexp will be generous in what sexps it accepts, while V2.t_of_sexp should not be expected to parse sexps not generated by V2.sexp_of_t

module With_t_of_sexp_abs : sig ... end

Provides a sexp representation where all sexps must include a timezone (in contrast to V1 above, which will assume local timezone if it's not specified). When serializing, it uses the local timezone. Concretely, this is just V1 but t_of_sexp is replaced with t_of_sexp_abs.

module Span : sig ... end

module Ofday : sig ... end

module Zone : sig ... end