Module Address.Unix

type t = [

| `Unix of string

]

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val __t_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> t

val compare : t -> t -> int

val create : string -> t

val to_string : t -> string

val to_sockaddr : t -> Core.Unix.sockaddr