Module Tcp.Where_to_connect

A Where_to_connect describes the socket that a tcp client should connect to.

type 'a t constraint 'a = [< Async_unix__Unix_syscalls.Socket.Address.t ]

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

type inet = [ `Inet of Unix.inet_addr * int ] t

val sexp_of_inet : inet -> Ppx_sexp_conv_lib.Sexp.t

type unix = [ `Unix of string ] t

val sexp_of_unix : unix -> Ppx_sexp_conv_lib.Sexp.t

val remote_address : 'a t -> 'a Async_kernel.Deferred.t

val of_host_and_port : ?⁠bind_to_address:Core.Unix.Inet_addr.t -> ?⁠bind_to_port:int -> Core.Host_and_port.t -> inet

bind_to_address and bind_to_port can be used to bind the source IP and port of the underlying socket. This does not necessarily alter the interface used to send the data. In particular, the commonly used destination-based routing is unaffected by binding to a different address.

val of_inet_address : ?⁠bind_to_address:Core.Unix.Inet_addr.t -> ?⁠bind_to_port:int -> [ `Inet of Unix.inet_addr * int ] -> inet

val of_file : string -> unix

val of_unix_address : [ `Unix of string ] -> unix