Module Unix.Inet_addr

include module type of struct include Core_unix.Inet_addr end

type t = Core_unix__.Import.Unix.inet_addr

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

val arg_type : t Core.Command.Arg_type.t

val t_of_sexp : Core.Sexp.t -> t

t_of_sexp is deprecated because it used to block to do a DNS lookup, and we don't want a sexp converter to do that. As we transition away, one can use Blocking_sexp, which has the old behavior.

module Blocking_sexp = Core_unix.Inet_addr.Blocking_sexp

Blocking_sexp performs DNS lookup to resolve hostnames to IP addresses.

include Core.Comparable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness = Core_unix.Inet_addr.comparator_witness

val comparator : ( t, comparator_witness ) Base.Comparator.comparator

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare = Core_unix.Inet_addr.Replace_polymorphic_compare

module Map = Core_unix.Inet_addr.Map

module Set = Core_unix.Inet_addr.Set

val of_string : string -> t

Conversion from the printable representation of an Internet address to its internal representation. The argument string consists of 4 numbers separated by periods (XXX.YYY.ZZZ.TTT) for IPv4 addresses, and up to 8 numbers separated by colons for IPv6 addresses. Raise Failure when given a string that does not match these formats.

val to_string : t -> string

Return the printable representation of the given Internet address. See of_string for a description of the printable representation.

val bind_any : t

A special address, for use only with bind, representing all the Internet addresses that the host machine possesses.

val bind_any_inet6 : t

val localhost : t

Special addresses representing the host machine.

127.0.0.1

val localhost_inet6 : t

(::1)

val inet4_addr_of_int32 : Core.Int32.t -> t

Some things (like the kernel) report addresses as hex or decimal strings. Provide conversion functions.

val inet4_addr_to_int32_exn : t -> Core.Int32.t

inet4_addr_to_int32_exn t = 0l when t = Inet_addr.of_string ("0.0.0.0"). An exception is raised if t is a not an IPv4 address.

val inet4_addr_of_int63 : Core.Int63.t -> t

val inet4_addr_to_int63_exn : t -> Core.Int63.t

module Stable = Core_unix.Inet_addr.Stable

val of_string_or_getbyname : string -> t Async_kernel.Deferred.t

of_string_or_getbyname hostname does a DNS lookup of hostname and returns the resulting IP address.