Module Unix.Name_info

type t = Core_unix.name_info = {

ni_hostname : string;

ni_service : string;

}

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

type getnameinfo_option = Core_unix.getnameinfo_option =

| NI_NOFQDN

| NI_NUMERICHOST

| NI_NAMEREQD

| NI_NUMERICSERV

| NI_DGRAM

val bin_shape_getnameinfo_option : Bin_prot.Shape.t

val bin_size_getnameinfo_option : getnameinfo_option Bin_prot.Size.sizer

val bin_write_getnameinfo_option : getnameinfo_option Bin_prot.Write.writer

val bin_writer_getnameinfo_option : getnameinfo_option Bin_prot.Type_class.writer

val bin_read_getnameinfo_option : getnameinfo_option Bin_prot.Read.reader

val __bin_read_getnameinfo_option__ : ( int -> getnameinfo_option ) Bin_prot.Read.reader

val bin_reader_getnameinfo_option : getnameinfo_option Bin_prot.Type_class.reader

val bin_getnameinfo_option : getnameinfo_option Bin_prot.Type_class.t

val sexp_of_getnameinfo_option : getnameinfo_option -> Sexplib0.Sexp.t

val getnameinfo_option_of_sexp : Sexplib0.Sexp.t -> getnameinfo_option

val get : sockaddr -> getnameinfo_option list -> t Async_kernel.Deferred.t