Module Writer.Flush_result

type t =

| Error	(*
Error is accompanied by a detailed error being sent to the writer's monitor.
*)

| Consumer_left	(*
Consumer_left is returned when the consumer leaves (see consumer_left) and raise_when_consumer_leaves is set to false. If that flag is set to true, then you get an Error instead.
*)

| Flushed of Time_ns_unix.t	(*
The time just after the write() system call returned or the time flushed_* was called if all the writes were already flushed by then.
*)

val sexp_of_t : t -> Sexplib0.Sexp.t