Module Async_unix__Async_sys

This module overrides everything in the Core.Sys module that might block. Functions do the same thing as their counterparts in Core.Sys, but instead return deferreds. For a description of their semantics see the documentation for the Core.Sys module.