Module Async_unix__Dump_core_on_job_delay

Dump core if jobs are delayed, to get additional debug information when running on UNIX systems that support core dumps.

It is not normally enabled, but may be enabled for any program by setting the appropriate field, dump_core_on_job_delay, in the ASYNC_CONFIG environment variable.