Module Async_unix__Fd

An Fd.t is a wrapper around a Unix file descriptor, with additional information about the kind of file descriptor and logic to ensure that we don't use a file descriptor that has been closed, or close a file descriptor that is in use.

Since Async uses multiple threads to make read/write and other system calls on file descriptors, and Unix reuses descriptors after they are closed, Async has to be very careful that the file descriptor passed to a system call is referring to the file it intends, and not some other completely unrelated file that Unix has decided to assign to the same descriptor.

Provided that one only accesses a file descriptor within the context of the functions below, Fd guarantees that the file descriptor will not have been closed/reused and will correspond to the same file that it did when the Fd.t was created:

      with_file_descr
      with_file_descr_deferred
      syscall
      syscall_exn
      syscall_result_exn
      syscall_in_thread
      syscall_in_thread_exn

The Fd module keeps track of which of these functions are currently accessing the file descriptor, and ensures that any close happens after they complete. Also, once close has been called, it refuses to provide further access to the file descriptor, either by returning a variant `Already_closed or by raising an exception.

Some of the above functions take an optional ?nonblocking:bool argument. The default is false, but if it is set to true, then before supplying the underlying file_descr, the Fd module will first call Unix.set_nonblock file_descr, if it hasn't previously done so on that file descriptor. This is intended to support making nonblocking system calls (e.g., connect, read, write) directly within Async, without releasing the OCaml lock or the Async lock, and without using another thread.