Module Async_unix__Reader

Reader is Async's main API for buffered input from a file descriptor. It is the analog of Stdio.In_channel.

Each reader has an internal buffer, which is filled via read() system calls when data is needed to satisfy a Reader.read* call.

Each of the read functions returns a deferred that will become determined when the read completes. It is an error to have two simultaneous reads. That is, if you call a read function, you should not call another read function until the first one completes.

If the file descriptor underlying a reader is closed, the reader will return EOF (after all the buffered bytes have been read).

Any Reader.read* call could, rather than determine its result, send an exception to the monitor in effect when read was called. Such exceptions can be handled in the usual way by using try_with, e.g.:

try_with (fun () -> Reader.read reader ...)