Module Async_unix__Tcp

Tcp supports connection to inet sockets and unix sockets.

These are two different types. We use 'a Where_to_connect.t to specify a socket to connect to, where the 'a identifies the type of socket.