Module Async_unix__Thread_safe

The Thread_safe module has functions that are safe to call from threads outside Async, such as the ones spawned by In_thread.run.

This is in contrast with the rest of Async library which is generally not thread-safe.

All the Thread_safe.block* and Thread_safe.run* functions wake up the Async scheduler to ensure that it continues in a timely manner with whatever jobs got started. Some functions take an optional ?wakeup_scheduler:bool argument, which defaults to true. One can cause the scheduler to not be woken up by supplying ~wakeup_scheduler:false, which can reduce CPU use, but increase latency, because the scheduler may not wake up for a while to process jobs.