Module Async_unix__Unix_syscalls

Unix_syscalls provides an interface to many of the functions in OCaml's standard Unix module. It uses a deferred in the return type of functions that would block. The idea is that in an Async program one does not use the standard Unix module, since in doing so one could accidentally block the whole program.

There are also a number of cosmetic changes (e.g., polymorphic variants) and other improvements (e.g., phantom types on sockets) over the standard Unix module.