Module Async_unix__Writer

Writer is Async's main API for output to a file descriptor. It is the analog of Core.Out_channel.

Each writer has an internal buffer, to which Writer.write* adds data. Each writer uses an Async cooperative thread that makes write() system calls to move the data from the writer's buffer to an OS buffer via the file descriptor.

There is no guarantee that the data sync on the other side of the writer can keep up with the rate at which you are writing. If it cannot, the OS buffer will fill up and the writer's cooperative thread will be unable to send any bytes. In that case, calls to Writer.write* will grow the writer's buffer without bound, as long as your program produces data. One solution to this problem is to call Writer.flushed and not continue until that becomes determined, which will only happen once the bytes in the writer's buffer have been successfully transferred to the OS buffer. Another solution is to check Writer.bytes_to_write and not produce any more data if that is beyond some bound.

There are two kinds of errors that one can handle with writers. First, a writer can be closed, which will cause future writes (and other operations) to synchronously raise an exception. Second, the writer's cooperative thread can fail due to a write() system call failing. This will cause an exception to be sent to the writer's monitor, which will be a child of the monitor in effect when the writer is created. One can deal with such asynchronous exceptions in the usual way, by handling the stream returned by Monitor.detach_and_get_error_stream (Writer.monitor writer).