Module Thread_pool.Helper_thread

A helper thread is a thread with its own dedicated work queue. Work added for the helper thread is guaranteed to be run by that thread. The helper thread only runs work explicitly supplied to it. Helper threads count towards a thread pool's max_num_threads.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val default_name : t -> string

default_name t returns the name that will be used for work performed by t, unless that work is added with an overriding name

val default_priority : t -> Priority.t

default_priority t returns the priority that will be used for work performed by t, unless that work is added with an overriding priority.