Module Address.Inet

type t = [

| `Inet of Unix.inet_addr * int

]

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : Core.Sexp.t -> t

val __t_of_sexp__ : Core.Sexp.t -> t

module Blocking_sexp : sig ... end

Blocking_sexp performs DNS lookup to resolve hostnames to IP addresses.

module Show_port_in_test : sig ... end

Show_port_in_test renders the port as an integer, even in tests, unlike the normal sexp_of_t and to_string, which render the port as "PORT" in tests.

val create : Unix.inet_addr -> port:int -> t

val create_bind_any : port:int -> t

val addr : t -> Unix.inet_addr

val port : t -> int

val to_string : t -> string

val to_host_and_port : t -> Core.Host_and_port.t

val to_sockaddr : t -> Core.Unix.sockaddr