Module Unix.Addr_info

type t = Core.Unix.addr_info = {

ai_family : socket_domain;

ai_socktype : socket_type;

ai_protocol : int;

ai_addr : sockaddr;

ai_canonname : string;

}

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : Core.Sexp.t -> t

module Blocking_sexp : sig ... end

Blocking_sexp performs DNS lookup to resolve hostnames to IP addresses.

type getaddrinfo_option = Core.Unix.getaddrinfo_option =

| AI_FAMILY of socket_domain

| AI_SOCKTYPE of socket_type

| AI_PROTOCOL of int

| AI_NUMERICHOST

| AI_CANONNAME

| AI_PASSIVE

val bin_shape_getaddrinfo_option : Bin_prot.Shape.t

val bin_size_getaddrinfo_option : getaddrinfo_option Bin_prot.Size.sizer

val bin_write_getaddrinfo_option : getaddrinfo_option Bin_prot.Write.writer

val bin_writer_getaddrinfo_option : getaddrinfo_option Bin_prot.Type_class.writer

val bin_read_getaddrinfo_option : getaddrinfo_option Bin_prot.Read.reader

val __bin_read_getaddrinfo_option__ : (int -> getaddrinfo_option) Bin_prot.Read.reader

val bin_reader_getaddrinfo_option : getaddrinfo_option Bin_prot.Type_class.reader

val bin_getaddrinfo_option : getaddrinfo_option Bin_prot.Type_class.t

val sexp_of_getaddrinfo_option : getaddrinfo_option -> Ppx_sexp_conv_lib.Sexp.t

val getaddrinfo_option_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> getaddrinfo_option

val get : ?⁠service:string -> host:string -> getaddrinfo_option list -> t list Async_kernel.Deferred.t