Module Unix.Stats

type t = {

dev : int;

ino : int;

kind : File_kind.t;

perm : file_perm;

nlink : int;

uid : int;

gid : int;

rdev : int;

size : int64;

atime : Core.Time.t;

mtime : Core.Time.t;

ctime : Core.Time.t;

}

val ctime : t -> Core.Time.t

val mtime : t -> Core.Time.t

val atime : t -> Core.Time.t

val size : t -> int64

val rdev : t -> int

val gid : t -> int

val uid : t -> int

val nlink : t -> int

val perm : t -> file_perm

val kind : t -> File_kind.t

val ino : t -> int

val dev : t -> int

module Fields : sig ... end

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

val of_unix : Core.Unix.stats -> t

val to_string : t -> string